pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.81.0",
    "feedparser>=6.0.12",
    "groq>=1.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
